Implement link lookups
Add an om.next-like syntax for specifying top-level links as an expression
inside queries. Works with both mg/pull and sg/pull (pass a nil lookup-ref or
use the new 2-arity variant).

Provide an easy api with mg/pull-link that takes a keyword instead of a
lookup-ref and will work as if calling pull with the ref at that key.

The subgraph api also implements this functionality, with a new self-registering
subscription handler at ::sg/pull-link.